1. News detail page and news list row Ui changes. 2. Moved screen package inside views package
