+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
             android:theme="@style/AppTheme"
             android:networkSecurityConfig="@xml/network_security_config"
             android:name=".NewsApp">
-        <activity android:name=".newslist.MainActivity">
+        <activity android:name="com.sachinsandbhor.newsapp.NewsActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <action android:name="android.intent.action.VIEW"/>
