+++ app/src/main/java/com/sachinsandbhor/newsapp/NewsActivity.kt
@@ -1,11 +1,11 @@
-package com.sachinsandbhor.newsapp.newslist
+package com.sachinsandbhor.newsapp
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.LifecycleOwner
 import com.sachinsandbhor.newsapp.R
 
-class MainActivity : AppCompatActivity(), LifecycleOwner {
+class NewsActivity : AppCompatActivity(), LifecycleOwner {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
