+++ app/src/main/java/com/sachinsandbhor/newsapp/entities/Article.kt
@@ -2,9 +2,10 @@ package com.sachinsandbhor.newsapp.entities
 
 import java.io.Serializable
 import android.widget.ImageView
+import android.widget.TextView
 import com.squareup.picasso.Picasso
 import androidx.databinding.BindingAdapter
-
+import com.sachinsandbhor.newsapp.utils.DateUtil
 
 
 data class Article(
@@ -22,4 +23,9 @@ fun loadImage(view: ImageView, imageUrl: String) {
     Picasso.get()
         .load(imageUrl)
         .into(view)
+}
+
+@BindingAdapter("app:dateText")
+fun dateToString(view: TextView, dateText: String) {
+    view.text = DateUtil.formatDate(dateText)
 }
