+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".newslist.MainActivity">
+        tools:context=".NewsActivity">
     <fragment
             android:name="androidx.navigation.fragment.NavHostFragment"
             android:layout_width="0dp"
