+++ app/src/main/res/layout/fragment_news_detail.xml
@@ -11,7 +11,7 @@
             android:layout_height="match_parent"
             android:background="@android:color/background_light"
             android:fitsSystemWindows="true"
-            tools:context=".newsdetail.NewsDetailFragment"
+            tools:context=".views.newsdetail.NewsDetailFragment"
     >
 
         <com.google.android.material.appbar.AppBarLayout
@@ -34,11 +34,12 @@
             >
 
                 <ImageView
-                        android:id="@+id/backdrop"
+                        android:id="@+id/news_backdrop_image"
                         android:layout_width="match_parent"
                         android:layout_height="match_parent"
                         android:scaleType="centerCrop"
                         android:fitsSystemWindows="true"
+                        android:contentDescription="@string/news_image_content_description"
                         app:imageUrl="@{article.urlToImage}"
                         app:layout_collapseMode="parallax"
                 />
@@ -60,49 +61,45 @@
         >
             <LinearLayout android:layout_width="match_parent" android:layout_height="match_parent"
                           android:orientation="vertical">
-                <androidx.constraintlayout.widget.ConstraintLayout
-                        android:layout_width="match_parent"
-                        android:layout_height="wrap_content">
 
-                    <TextView
-                            android:layout_width="0dp"
-                            android:layout_height="wrap_content"
-                            android:id="@+id/dateTextview" app:layout_constraintStart_toStartOf="parent"
-                            android:layout_marginLeft="8dp" android:layout_marginStart="8dp"
-                            android:layout_marginTop="16dp"
-                            android:text="@{article.publishedAt}"
-                            app:layout_constraintTop_toTopOf="parent" app:layout_constraintEnd_toEndOf="parent"
-                            android:layout_marginEnd="8dp" android:layout_marginRight="8dp"/>
-                    <TextView
-                            android:layout_width="0dp"
-                            android:layout_height="wrap_content"
-                            android:id="@+id/title" app:layout_constraintStart_toStartOf="parent"
-                            android:layout_marginLeft="8dp" android:layout_marginStart="8dp"
-                            android:layout_marginTop="8dp"
-                            app:layout_constraintTop_toBottomOf="@+id/dateTextview"
-                            app:layout_constraintEnd_toEndOf="parent" android:layout_marginEnd="8dp"
-                            android:text="@{article.title}"
-                            android:layout_marginRight="8dp"/>
-                    <TextView
-                            android:layout_width="0dp"
-                            android:layout_height="wrap_content"
-                            android:id="@+id/description" app:layout_constraintStart_toStartOf="parent"
-                            android:layout_marginLeft="8dp" android:layout_marginStart="8dp"
-                            app:layout_constraintEnd_toEndOf="parent" android:layout_marginEnd="8dp"
-                            android:layout_marginRight="8dp" android:layout_marginTop="8dp"
-                            android:text="@{article.description}"
-                            app:layout_constraintTop_toBottomOf="@+id/title"/>
-                    <TextView
-                            android:layout_width="0dp"
-                            android:layout_height="0dp"
-                            android:id="@+id/content" app:layout_constraintStart_toStartOf="parent"
-                            android:layout_marginLeft="8dp" android:layout_marginStart="8dp"
-                            app:layout_constraintEnd_toEndOf="parent" android:layout_marginEnd="8dp"
-                            android:layout_marginRight="8dp" android:layout_marginBottom="8dp"
-                            app:layout_constraintBottom_toBottomOf="parent" android:layout_marginTop="8dp"
-                            android:text="@{article.content}"
-                            app:layout_constraintTop_toBottomOf="@+id/description"/>
-                </androidx.constraintlayout.widget.ConstraintLayout>
+
+                <TextView
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:id="@+id/title"
+                        android:layout_marginLeft="8dp"
+                        android:layout_marginStart="8dp"
+                        android:layout_marginTop="8dp"
+                        android:layout_marginEnd="8dp"
+                        android:text="@{article.title}"
+                        android:layout_marginRight="8dp"
+                        android:textStyle="bold"
+                        android:textAppearance="@style/TextAppearance.MaterialComponents.Headline6"/>
+
+                <TextView
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:id="@+id/dateTextview"
+                        android:layout_marginLeft="8dp"
+                        android:layout_marginStart="8dp"
+                        android:layout_marginTop="16dp"
+                        app:dateText="@{article.publishedAt}"
+                        android:layout_marginEnd="8dp"
+                        android:layout_marginRight="8dp"
+                        android:textAppearance="@style/TextAppearance.MaterialComponents.Body2"/>
+
+                <TextView
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:id="@+id/content"
+                        android:layout_marginLeft="8dp"
+                        android:layout_marginStart="8dp"
+                        android:layout_marginEnd="8dp"
+                        android:layout_marginRight="8dp"
+                        android:layout_marginBottom="8dp"
+                        android:layout_marginTop="8dp"
+                        android:text="@{article.content}"
+                        android:textAppearance="@style/TextAppearance.MaterialComponents.Body2"/>
             </LinearLayout>
         </androidx.core.widget.NestedScrollView>
