+++ app/src/main/res/layout/fragment_news_list.xml
@@ -4,7 +4,7 @@
                                                    xmlns:tools="http://schemas.android.com/tools"
                                                    android:layout_width="match_parent"
                                                    android:layout_height="match_parent"
-                                                   tools:context=".newslist.NewsListFragment"
+                                                   tools:context=".views.newslist.NewsListFragment"
 >
     <ProgressBar
             style="?android:attr/progressBarStyle"
