+++ app/src/main/res/layout/news_item.xml
@@ -7,23 +7,28 @@
     <ImageView
             android:layout_width="100dp"
             android:layout_height="100dp" tools:srcCompat="@tools:sample/avatars"
-            android:id="@+id/newsimage"
+            android:id="@+id/news_thumbnail"
+            android:contentDescription="@string/news_image_content_description"
             app:layout_constraintStart_toStartOf="parent" android:layout_marginLeft="8dp"
-            android:layout_marginStart="8dp" android:layout_marginTop="8dp" app:layout_constraintTop_toTopOf="parent"/>
+            android:layout_marginStart="8dp" android:layout_marginTop="8dp" app:layout_constraintTop_toTopOf="parent"
+            android:layout_marginBottom="8dp" app:layout_constraintBottom_toBottomOf="parent"
+            android:scaleType="centerInside"/>
     <TextView
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:id="@+id/description" app:layout_constraintStart_toEndOf="@+id/newsimage"
+            android:id="@+id/news_description" app:layout_constraintStart_toEndOf="@+id/news_thumbnail"
             android:layout_marginLeft="8dp" android:layout_marginStart="8dp"
-            app:layout_constraintTop_toTopOf="@+id/newsimage" app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
             android:layout_marginEnd="8dp" android:layout_marginRight="8dp"
-            android:textAppearance="@style/TextAppearance.AppCompat.Body1" />
+            android:textAppearance="@style/TextAppearance.MaterialComponents.Body1"
+            app:layout_constraintTop_toTopOf="@+id/news_thumbnail"/>
     <TextView
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:id="@+id/datetextview" app:layout_constraintStart_toEndOf="@+id/newsimage"
+            android:textAppearance="@style/TextAppearance.MaterialComponents.Body2"
+            android:id="@+id/news_date" app:layout_constraintStart_toEndOf="@+id/news_thumbnail"
             android:layout_marginLeft="8dp" android:layout_marginStart="8dp" app:layout_constraintEnd_toEndOf="parent"
             android:layout_marginEnd="8dp" android:layout_marginRight="8dp" android:layout_marginTop="8dp"
-            app:layout_constraintTop_toBottomOf="@+id/description" android:gravity="end"
+            app:layout_constraintTop_toBottomOf="@+id/news_description" android:gravity="end"
     />
 </androidx.constraintlayout.widget.ConstraintLayout>
