+++ app/src/main/res/navigation/news_nav_graph.xml
@@ -5,7 +5,7 @@
             app:startDestination="@id/newsListFragment">
 
     <fragment android:id="@+id/newsListFragment"
-              android:name="com.sachinsandbhor.newsapp.newslist.NewsListFragment"
+              android:name="com.sachinsandbhor.newsapp.views.newslist.NewsListFragment"
               android:label="fragment_news_list"
               tools:layout="@layout/fragment_news_list">
 
@@ -14,7 +14,7 @@
     </fragment>
 
     <fragment android:id="@+id/newsDetailFragment"
-              android:name="com.sachinsandbhor.newsapp.newsdetail.NewsDetailFragment"
+              android:name="com.sachinsandbhor.newsapp.views.newsdetail.NewsDetailFragment"
               android:label="fragment_news_detail"
               tools:layout="@layout/fragment_news_detail">
         <argument android:name="article" app:argType="com.sachinsandbhor.newsapp.entities.Article"/>
